refactor: trim type declaration
