+++ src/components/TravelProduct/TravelProduct.tsx
@@ -1,19 +1,8 @@
 import classNames from 'classnames';
-import { ItemSize } from '../Carousel/Carousel';
+
 import styles from './TravelProduct.module.css';
 
-type SeatType = '일반석' | '이등석' | '일등석';
-type BookingType = '편도' | '왕복';
-export type TravelProductData = Readonly<{
-  id: number;
-  departure: string;
-  destination: string;
-  seatType: SeatType;
-  bookingType: BookingType;
-  minimumFare: number;
-  imageUrl: string;
-  href: string;
-}>;
+import { ItemSize, TravelProductData } from '../../types';
 
 type TravelProductType = TravelProductData & {
   itemSize: ItemSize;
